requirements.txt
google-generativeai
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai